requirements.txt
GitPython
netifaces
bluepy
paho-mqtt
python-eq3bt
voluptuous
homeassistant
